-chatserver will now stop spamming heartbeats some hundred times a second -chatserver will now use chatserver.lpg to log events -database was regularly crashing the servers when an (erroneous) query result returned no result - this has been fixed by not handling the return of such a query. This might lead to orphaned async containers however so its no excuse to keep queries broken. A proper log message will be displayed -added the infrastructure for auto delta and baseline messages. -transferred the code for auto packet creation and the actual message structures from the new core -some more comentation in the code -removed some no longer necessary debu code in the mesagefactory and updated comments and log messages -removed some more instances of BString -removed eruptors heightmap code -fixed an issue in PKs terrain service where working_fractal_ might not be defined if a scene (like tutorial) has no terrain -implemented heightmap lookups through PKs TRE viewer and Terrain service -removed sevveral unnecessary includes out of playerobject.h

+++ src/ChatServer/ChatServer.h
@@ -96,7 +96,7 @@ private:
     CSRManager*					  mCSRManager;
 
     DispatchClient*				  mClient;
-    uint32					      mLastHeartbeat;
+    uint64					      mLastHeartbeat;
 
 };
